Updating with changes done for the 7.1 release

+++ books/centaur/quicklisp/clean.sh
@@ -32,19 +32,16 @@
 
 set -e
 
-# Modification by Matt K. to go with ACL2 release: Comment out lines
-# that assume we are in a git repository.
+DIFF=`git status --porcelain bundle`
+if [ ! -z "$DIFF" ]
+then
+    echo "Cowardly refusing to clean because there are uncommitted changes in bundle/"
+    git status bundle | sed 's/^/   | /'
+    exit 1
+fi
 
-# DIFF=`git status --porcelain bundle`
-# if [ ! -z "$DIFF" ]
-# then
-#     echo "Cowardly refusing to clean because there are uncommitted changes in bundle/"
-#     git status bundle | sed 's/^/   | /'
-#     exit 1
-# fi
-# 
-# echo "Cleaning quicklisp/bundle"
-# git clean -f bundle
+echo "Cleaning quicklisp/bundle"
+git clean -f bundle
 
 echo "Cleaning asdf-home/cache/common-lisp"
 rm -rf asdf-home/cache/common-lisp
